% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beautifytable.R
\name{beautifytable}
\alias{beautifytable}
\title{beautifytable}
\usage{
beautifytable(.data, caption, size = 7, width = "3in")
}
\arguments{
\item{.data}{The dataframe or table which is being beautified}

\item{caption}{The caption for the table}

\item{size}{The font size. Default size is 7.}

\item{width}{The width of the first column of the table. Default size is 3 inches.}
}
\value{
Output in Rmarkdown following pdf knit
}
\description{
This function beautifies tables for Rmarkdown pdf outputs
}
\examples{
irisflowers <- iris \%>\%
  group_by(Species) \%>\%
  summarise(n = n())
beautifytable(irisflowers, "Flower Count")
}
