% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corsig.R
\name{corsig}
\alias{corsig}
\title{corsig}
\usage{
corsig(.data, decimals = 2, method = "pearson", numbered = T)
}
\arguments{
\item{.data}{a dataframe or matrix with numeric values}

\item{decimals}{the number of decimals that are shown in the correlations}

\item{method}{methods used based on stats::cor methods}

\item{numbered}{whether to number the row and column names to condense results}
}
\value{
a dataframe with correlations and significance based on p-values
}
\description{
This function follows stats::cor results to provide correlation results and
significance based on p-values in a table
}
\examples{
irisnum <- iris \%>\%
  select_if(is.numeric)
corsig(irisnum, decimals = 3, method = "kendall", numbered = T)
corsig(irisnum, decimals = 2, method = "pearson", numbered = F)



}
