% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasummariser.R
\docType{package}
\name{datasummariser}
\alias{datasummariser}
\title{datasummariser: A package for summarising and analysing data into nice formats}
\description{
The datasummariser package current provides two important functions:
}
\details{
Tidy en-mass analyses and determining issues within data
}
\section{datasummariser functions}{

mass_ttest and thresholdchecker
}

\seealso{
\code{\link{thresholdcheck}} and \code{\link{mass_ttest}}
}
