% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mass_ttest.R
\name{mass_ttest}
\alias{mass_ttest}
\title{Mass T-Test}
\usage{
mass_ttest(.data, x, y, ...)
}
\arguments{
\item{.data}{dataframe that we're working with.}

\item{x}{grouping variable with two levels}

\item{y}{all numerical variables being compared by x}

\item{...}{all parameters from t.test from the psych package}
}
\value{
A tidy table of t-test results with Cohen's D on all variables in \code{varlist} between two groups in \code{variable}
}
\description{
Runs t-test results between two groups for multiple variables
}
\examples{
mass_ttest(mtcars, vs, c("cyl", "mpg"))
mass_ttest(mtcars, vs, c("cyl", "mpg"), var.equal = TRUE)

}
\seealso{
\code{\link{thresholdcheck}}
}
