% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thresholdcheck.R
\name{thresholdcheck}
\alias{thresholdcheck}
\title{Threshold Check}
\usage{
thresholdcheck(.data, variable, vector)
}
\arguments{
\item{.data}{dataframe which contains variable of interest}

\item{variable}{variable which is being tested for thresholds}

\item{vector}{vector containing expected values}
}
\value{
If there are rows with values outside of the expected values, a dataframe with these rows will appear.
}
\description{
This function tests whether there are unexpected values for a variable in a dataframe.
}
\examples{
thresholdcheck(iris, Species, c("setosa"))
thresholdcheck(mtcars, cyl, 1:5)

}
\seealso{
\code{\link{mass_ttest}}
}
