% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasummariser.R
\docType{package}
\name{datasummariser}
\alias{datasummariser}
\title{datasummariser: A package for summarising and analysing data into nice formats}
\description{
The datasummariser package current provides two important functions:
}
\details{
1. Tidy en-mass analyses with p-value significance results
2. Producing well-formatted tables when knitting to PDF in Rmarkdown
}
\section{datasummariser functions}{

}

\seealso{
\code{\link{mass_ttest}}

\code{\link{thresholdcheck}}

\code{\link{efasummariser}}

\code{\link{corsig}}

\code{\link{beautifytable}}
}
